files saved

+++ pandemic-management-app/src/app/listuser/listuser.component.ts
@@ -51,11 +51,15 @@ export class ListuserComponent implements OnInit {
       
       }
       delete(id:any,rev:any){
-        this.api.Delete(id,rev).subscribe(res=>{
-          console.log(res);
-        alert("Deleted sucessfully");
-        window. location. reload();
-        })
+        if (confirm("Do you really want to delete ?") === true) {
+          this.api.Delete(id,rev).subscribe(res=>{
+            console.log(res);
+            window. location. reload();
+  
+          })
+        } else {
+          window. location. reload();
+        }
       }
       }
